Yank attendance stuff to its own AttendanceController from EventController
