+++ src/controllers/index.ts
@@ -4,6 +4,7 @@ import { AuthController, AuthControllerImpl } from './AuthController';
 import { TeapotController, TeapotControllerImpl } from './TeapotController';
 import { InductionClassController, InductionClassControllerImpl } from './InductionClassController';
 import { PointsController, PointsControllerImpl } from './PointsController';
+import { AttendanceController, AttendanceControllerImpl } from './AttendanceController';
 
 export const controllers = [
   EventController,
@@ -12,6 +13,7 @@ export const controllers = [
   TeapotController,
   InductionClassController,
   PointsController,
+  AttendanceController,
 ];
 
 // map from name of controller to an instance
@@ -23,6 +25,7 @@ controllerMap.set(AuthController.name, AuthControllerImpl);
 controllerMap.set(TeapotController.name, TeapotControllerImpl);
 controllerMap.set(InductionClassController.name, InductionClassControllerImpl);
 controllerMap.set(PointsController.name, PointsControllerImpl);
+controllerMap.set(AttendanceController.name, AttendanceControllerImpl);
 
 export const ControllerContainer = {
   // controller here is just a class and all classes have names in ts
