+++ src/payloads/index.ts
@@ -1,4 +1,5 @@
 export {
+  EventIDRequest,
   EventRequest,
   EventResponse,
   MultipleEventResponse,
